Test : Float in zbuffer

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: vdurand <vdurand@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/21 10:22:29 by halnuma           #+#    #+#             */
-/*   Updated: 2025/09/07 20:16:38 by vdurand          ###   ########.fr       */
+/*   Updated: 2025/09/08 16:51:00 by vdurand          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -148,7 +148,7 @@ typedef struct s_game
 	t_png_pixel8		f_color;
 	t_png_pixel8		c_color;
 	t_tilemap			*tilemap;
-	uint8_t				*z_buffer;
+	float				*z_buffer;
 	t_key				key_buffer[KEY_MAX_COUNT];
 	t_png				*textures[TEXTURE_MAX_COUNT];
 	t_game_state		state;
